refactor: split out `filterDraftEntities` in a function
